Add 'recipes' directory and 'casserole.html'
Created a new directory where all pages besides index will be stored
for file structure. Created a new page called casserole.html and added
basic HTML boilerplate and a H1 heading for a title.
This commit is part of the initial setup for organizing recipe files
and the first recipes page.